requirements.txt
# 《是男人就砍一刀》游戏基础依赖包
# 包含游戏运行所需的核心依赖

# 游戏引擎
pygame>=2.5.0,<3.0.0

# HTTP请求库
requests>=2.31.0,<3.0.0

# 环境变量管理
python-dotenv>=1.0.0,<2.0.0

# 类型注解扩展
typing-extensions>=4.7.0,<5.0.0

# 日志增强（可选）
colorlog>=6.7.0,<8.0.0

# 数据处理
pandas>=2.0.0,<3.0.0

# JSON处理增强
ujson>=5.8.0,<6.0.0

# 时间处理
python-dateutil>=2.8.0,<3.0.0

# 数学计算增强
numpy>=1.24.0,<2.0.0

# 文件路径处理
pathlib2>=2.3.0,<3.0.0; python_version<"3.4"

# 配置文件处理
pyyaml>=6.0,<7.0.0

# 缓存支持
diskcache>=5.6.0,<6.0.0

# 异步支持（可选，用于未来扩展）
aiohttp>=3.8.0,<4.0.0

# 加密支持（用于存档等）
cryptography>=41.0.0,<42.0.0